KAFKA-8021 Fix unexpected behavior in KafkaProducer.flush() during a ProducerBatch is split
flush() waits on produce batch future. Incase when the batch splits it creates
new batches and before these batches are added to incomplete batch list
the big batch future is marked as completed.
Between the marking of future as compelete and adding the new batches into incomplete list
if flush() is called it'll only wait for existing complete batches not including
the newly split batches.

The fix is to delay the marking of big batch as completed until the new batches are added
into the incomplete list.